requirements.txt
pycryptodome==3.16.0
pytest==7.2.0
requests==2.25.1
